coreos: Move files/ into installers/coreos
These packages are only used by coreso (aka flatcar) so move them there.
