+++ installers/coreos/files/ignition/config.go
