+++ installers/coreos/files/ignition/storage.go
